fix invalid: malformed module path
