--- gin/v0/go.mod
@@ -1,8 +1,0 @@
-module github.com/SkyAPM/go2sky-plugins/gin/v0
-
-go 1.12
-
-require (
-	github.com/SkyAPM/go2sky v0.4.0
-	github.com/gin-gonic/gin v1.3.0
-)
