+++ gin/v2/go.mod
@@ -4,5 +4,5 @@ go 1.12
 
 require (
 	github.com/SkyAPM/go2sky v0.4.0
-	github.com/gin-gonic/gin v1.5.0
+	github.com/gin-gonic/gin v1.3.0
 )
