+++ gin/v3/example_gin_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package v0
+package v3
 
 import (
 	"fmt"
