+++ gin/v3/go.mod
@@ -1,0 +1,8 @@
+module github.com/SkyAPM/go2sky-plugins/gin/v3
+
+go 1.12
+
+require (
+	github.com/SkyAPM/go2sky v0.4.0
+	github.com/gin-gonic/gin v1.5.0
+)
